+++ easyaccess/config_ea.py
@@ -23,7 +23,7 @@ except NameError:
 configcomment = """#
 # Easyaccess default parameters
 #
-# database        : Default is dessci, change to desoper, destest, oldoper, desdr and others
+# database        : Default is dessci, change to desoper, destest, desdr and others
 #                   Make sure the db-"database" section is in the .desservices.ini
 # editor          : Default editor to open from inside easyaccess if $EDITOR is not set
 # prefetch        : Prefetch number of rows to get from oracle (not the number of total rows)
@@ -162,7 +162,6 @@ def get_desconfig(desfile, db, verbose=True, user=None, pw1=None):
     """
     Loads des config file or create one if it does not exist.
     """
-    server_n = 'leovip148.ncsa.uiuc.edu'
     server_2 = 'desdb.ncsa.illinois.edu'
     server_3 = 'desdb-dr.ncsa.illinois.edu'
     port_n = '1521'
@@ -170,7 +169,6 @@ def get_desconfig(desfile, db, verbose=True, user=None, pw1=None):
     if not db[:3] == 'db-':
         db = 'db-' + db
 
-    db_alter = None
     config = configparser.ConfigParser()
     configwrite = False
     check = config.read(desfile)
@@ -183,11 +181,11 @@ def get_desconfig(desfile, db, verbose=True, user=None, pw1=None):
         if verbose:
             print()
 
-    databases = ['db-desoper', 'db-dessci', 'db-destest', 'db-oldoper', 'db-desdr']
+    databases = ['db-desoper', 'db-dessci', 'db-destest', 'db-desdr']
 
     if db not in databases and not config.has_section(db):
         check_db = input(
-            '\nDB entered is not dessci, desoper, destest,  oldoper or desdr '
+            '\nDB entered is not dessci, desoper, destest or desdr '
             'or in DES_SERVICE file, continue anyway [y]/n\n')
         if check_db in ('n', 'N', 'no', 'No', 'NO', 'nO'):
             sys.exit(0)
@@ -198,14 +196,14 @@ def get_desconfig(desfile, db, verbose=True, user=None, pw1=None):
         configwrite = True
         if db == 'db-dessci':
             kwargs = {'host': server_2, 'port': port_n, 'service_name': 'dessci'}
-        elif db == 'db-oldoper':
-            kwargs = {'host': server_n, 'port': port_n, 'service_name': 'desoper'}
         elif db == 'db-desdr':
             kwargs = {'host': server_3, 'port': port_n, 'service_name': 'desdr'}
+        elif db == 'db-destest':
+            kwargs = {'host': server_2, 'port': port_n, 'service_name': 'destest'}
         elif db == 'db-desoper':
             kwargs = {'host': server_2, 'port': port_n, 'service_name': 'desoper'}
         else:
-            kwargs = {'host': server_n, 'port': port_n, 'service_name': db[3:]}
+            kwargs = {'host': server_2, 'port': port_n, 'service_name': db[3:]}
         dsn = cx_Oracle.makedsn(**kwargs)
         good = False
         if user is None:
@@ -274,13 +272,6 @@ def get_desconfig(desfile, db, verbose=True, user=None, pw1=None):
             if not config.has_option(db, 'server'):
                 configwrite = True
                 config.set(db, 'server', server_2)
-    elif db == 'db-oldoper':
-        if not config.has_option(db, 'name'):
-            configwrite = True
-            config.set(db, 'name', 'desoper')
-            if not config.has_option(db, 'server'):
-                configwrite = True
-                config.set(db, 'server', server_n)
     elif db == 'db-desdr':
         if not config.has_option(db, 'name'):
             configwrite = True
@@ -294,7 +285,7 @@ def get_desconfig(desfile, db, verbose=True, user=None, pw1=None):
             config.set(db, 'name', db[3:])
             if not config.has_option(db, 'server'):
                 configwrite = True
-                config.set(db, 'server', server_n)
+                config.set(db, 'server', server_2)
     if not config.has_option(db, 'port'):
         configwrite = True
         config.set(db, 'port', port_n)